ADD: change address

+++ service/user/address.js
@@ -6,31 +6,30 @@ module.exports = {
     return new Promise(async (res, rej) => {
       try {
         data["user_id"] = _id;
-        let userData = [await userModel.findById(_id)];
-        if (userData) {
+        let userData = await userModel.find({ _id: _id });
+        console.log("userData ....", userData.length);
+        if (userData.length > 0) {
           data.email = userData[0].email;
           data.mobile = userData[0].mobile;
-          let existData = [await addressModel.findOne({ user_id: _id })];
-          // console.log("existData.length ...........", existData.length);
+          let existData = await addressModel.find({ user_id: _id });
+          console.log("existData.length ...........", existData.length);
           if (existData.length > 0) {
-            let removedData = await addressModel.findOneAndDelete({
-              user_id: _id,
-            });
-            let ifNnewAddressModel = new addressModel(data);
-            let if_saveData = await ifNnewAddressModel.save();
-            if (if_saveData) {
-              res({ status: 200, data: "Data Updated Successfully!!" });
-            } else {
-              rej({ status: 404, message: "something went wrong!!" });
-            }
+            let removedData = await addressModel.findOneAndDelete({ user_id: _id });
+            //   let ifNnewAddressModel = new addressModel(data);
+            //   let if_saveData = await ifNnewAddressModel.save();
+            //   if (if_saveData) {
+            //     res({ status: 200, data: "Data Added Successfully!!" });
+            //   } else {
+            //     rej({ status: 404, message: "something went wrong!!" });
+            //   }
+            // } else {
+          }
+          let newAddressModel = new addressModel(data);
+          let saveData = await newAddressModel.save();
+          if (saveData) {
+            res({ status: 200, data: "Data Added Successfully!!" });
           } else {
-            let newAddressModel = new addressModel(data);
-            let saveData = await newAddressModel.save();
-            if (saveData) {
-              res({ status: 200, data: "Data Added Successfully!!" });
-            } else {
-              rej({ status: 404, message: "something went wrong!!" });
-            }
+            rej({ status: 404, message: "something went wrong!!" });
           }
         } else {
           rej({ status: 404, message: "user not found!!" });
@@ -98,7 +97,7 @@ module.exports = {
       try {
         let updateData = await addressModel.findOneAndUpdate(
           { user_id: user_id, "address._id": address_id },
-          { $pull: { "address": {_id:address_id} } },
+          { $pull: { "address": { _id: address_id } } },
           { new: true }
         );
         if (updateData) {
